requirements.txt
reka-api
python-dotenv==1.0.0
rich==13.7.0
PyGithub==2.1.1
spotipy==2.23.0
